+++ maptasker/src/primitem.py
@@ -132,6 +132,13 @@ class PrimeItems:
         "Spanish": "es",
         "German": "de",
         "Simplified Chinese": "zh_CN",
+        "Traditional Chinese": "zh_TW",
+        "Hindi": "hi",
+        "French": "fr",
+        "Portuguese": "pt",
+        "Japanese": "ja",
+        "Russian": "ru",
+        "Korean": "ko",
     }
     languages_translated: ClassVar[dict[str, str]] = {}